backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.1
Deprecated==1.2.18
dnspython==2.6.1
ecdsa==0.19.1
elastic-transport==8.17.1
elasticsearch==8.0.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
iniconfig==2.1.0
Jinja2==3.1.6
limits==3.13.0
MarkupSafe==2.1.5
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2025.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.44.0
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
wrapt==1.17.2
zipp==3.20.2
